wayfinder-fox/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "wayfinder-fox"
version = "0.1.0"
description = "Fox (Waagosh) - Context enrichment for alert automation"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "contextcore-rabbit>=0.1.0",
    "contextcore>=0.1.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
kubernetes = ["kubernetes>=28.1.0"]
dev = ["pytest>=7.0", "mypy>=1.0", "ruff>=0.1"]

[tool.setuptools.packages.find]
where = ["src"]
